fix: properly implement `Status` interface
Removes the type declarations for the `GoogleAPIError` class' properties, as the types are inferred from the implemented interface, `Status`.

Omits the `message` and `errors` properties from the `Status` interface when implementing to the  `GoogleAPIError` class.

Copies the JSDoc descriptions from the `Status` interface to the `GoogleAPIError` class to fix the `eslint(jsdoc/require-jsdoc)` warning.