build: migrate from TypeScript's tsc to tsup
Removes tsc-alias and tslib as dependencies and removes all related options from tsconfig.json.

Adds tsup as a dependency and creates a configuration file for bundling the TypeScript files for production, as well as updating the module/module resolution and emitting related options in tsconfig.json to reflect the move to a bundler.

Changes the build script to use tsup instead of tsc.